requirements.txt
yt-dlp==2025.1.26
openai==1.61.1
pydub==0.25.1
python-dotenv==1.0.1
streamlit==1.42.0
ffmpeg-python==0.2.0
sentence-transformers==3.4.1
faiss-cpu==1.10.0
langchain==0.3.18
langchain-groq==0.2.4
langchain-huggingface==0.1.2
langchain-community==0.3.17